--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/DiffPane/components/DiffCodeViewHeader/index.ts
@@ -1,1 +1,0 @@
-export { DiffCodeViewHeader } from "./DiffCodeViewHeader";
