--- apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/DiffPane/components/DiffFileHeader/index.ts
@@ -1,1 +1,0 @@
-export { DiffFileHeader } from "./DiffFileHeader";
