+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/DiffPane/components/DiffHeaderMetadata/index.ts
@@ -1,0 +1,1 @@
+export { DiffHeaderMetadata } from "./DiffHeaderMetadata";
