+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/DiffPane/components/DiffHeaderPrefix/DiffHeaderPrefix.tsx
@@ -1,0 +1,43 @@
+import { ChevronDown, ChevronRight } from "lucide-react";
+import { useCallback } from "react";
+import { FileIcon } from "renderer/lib/fileIcons";
+import type { ChangesetFile } from "../../../../../useChangeset";
+
+interface DiffHeaderPrefixProps {
+	file: ChangesetFile;
+	collapsed: boolean;
+	onSetCollapsed: (path: string, value: boolean) => void;
+}
+
+export function DiffHeaderPrefix({
+	file,
+	collapsed,
+	onSetCollapsed,
+}: DiffHeaderPrefixProps) {
+	const onToggle = useCallback(
+		() => onSetCollapsed(file.path, !collapsed),
+		[onSetCollapsed, file.path, collapsed],
+	);
+
+	return (
+		// Flex row so the chevron button and FileIcon stay on one line —
+		// Tailwind's preflight forces `img { display: block }`, which would
+		// otherwise drop the FileIcon below the button inside Pierre's
+		// default block slot wrapper.
+		<div className="flex shrink-0 items-center gap-1">
+			<button
+				type="button"
+				onClick={onToggle}
+				aria-label={collapsed ? "Expand file" : "Collapse file"}
+				className="rounded p-1 text-muted-foreground/60 transition-colors hover:bg-accent hover:text-muted-foreground"
+			>
+				{collapsed ? (
+					<ChevronRight className="size-3.5" />
+				) : (
+					<ChevronDown className="size-3.5" />
+				)}
+			</button>
+			<FileIcon fileName={file.path} className="size-3.5 shrink-0" />
+		</div>
+	);
+}
